requirements.txt
pandas
numpy 
scikit-learn
requests
mysql-connector-python
python-dotenv
pyaml
dagshub
mlflow
xgboost
catboost

#-e .